chore: мигрировал систему с PosgreSQL на SQLite

+++ src/pybot/db/models/attachments.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sqlalchemy import BigInteger, ForeignKey, Text
+from sqlalchemy import ForeignKey, Integer, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ..base_class import Base
@@ -15,10 +15,10 @@ if TYPE_CHECKING:
 class Attachment(Base):
     __tablename__ = "attachments"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(Text, nullable=False)
     content: Mapped[str] = mapped_column(Text, nullable=False)
-    type_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("attachments_types.id"), nullable=False)
+    type_id: Mapped[int] = mapped_column(Integer, ForeignKey("attachments_types.id"), nullable=False)
 
     task_attachments: Mapped[list[TaskAttachment]] = relationship(
         "TaskAttachment",
